♻️🏢 Extract loader to dedicated file
