+++ apps/admin/src/routes/_layout.show.exhibitors.$id._index/card-description.tsx
@@ -8,7 +8,7 @@ import { ProfileStatusIcon } from "#show/exhibitors/profile/status";
 import { ExhibitorStatus } from "#show/exhibitors/status";
 import { StatusHelper } from "#show/exhibitors/status-helper";
 import { useLoaderData } from "@remix-run/react";
-import type { loader } from "./route";
+import type { loader } from "./loader.server";
 
 export function CardDescription() {
   const { exhibitor } = useLoaderData<typeof loader>();
