+++ apps/admin/src/routes/_layout.show.exhibitors.$id._index/card-documents.tsx
@@ -9,7 +9,7 @@ import { DocumentsStatusIcon } from "#show/exhibitors/documents/status";
 import { ExhibitorStatus } from "#show/exhibitors/status";
 import { StatusHelper } from "#show/exhibitors/status-helper";
 import { useLoaderData } from "@remix-run/react";
-import type { loader } from "./route";
+import type { loader } from "./loader.server";
 
 export function CardDocuments() {
   const { exhibitor } = useLoaderData<typeof loader>();
